requirements.txt
python-dotenv==1.0.0
pymongo==4.6.1
google-generativeai==0.3.2
chromadb==0.4.22
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3